requirements.txt
Django==1.7.6
Markdown==2.6.1
SwampDragon==0.4.1.2
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
django-admin-bootstrapped==2.3.4
django-filter==0.9.2
djangorestframework==3.1.0
python-dateutil==2.4.1
redis==2.10.3
six==1.9.0
sockjs-tornado==1.0.1
tornado==4.1
tornado-redis==2.4.18
wsgiref==0.1.2
